+++ drag-and-drop/core/scripts/custom-script.ts
@@ -1,5 +1,0 @@
-// this is just an example script showing how you could add a custom script that is called in your packages
-// ak-postbuild script
-
-// eslint-disable-next-line no-console
-console.log('Done!');