+++ drag-and-drop/core/src/entry-point/util/set-custom-native-drag-preview.ts
@@ -1,0 +1,1 @@
+export { setCustomNativeDragPreview } from '../../util/set-custom-native-drag-preview';