+++ linking-platform/link-create/src/index.ts
@@ -1,4 +1,4 @@
 export type { LinkCreatePlugin } from './ui/types';
-export { default } from './ui/index';
+export { default, TextField, Select } from './ui/index';
 
 export { useLinkCreateCallback } from './controllers/callback-context/index';