+++ drag-and-drop/autoscroll/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/drag-and-drop-autoscroll
 
+## 0.1.13
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 0.1.12
 
 ### Patch Changes