+++ drag-and-drop/indicator/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/drag-and-drop-indicator
 
+## 0.9.1
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 0.9.0
 
 ### Minor Changes