+++ editor/adf-schema/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/adf-schema
 
+## 25.2.2
+
+### Patch Changes
+
+- [`8c04b73312e`](https://bitbucket.org/atlassian/atlassian-frontend/commits/8c04b73312e) - ED-16758 Added data-cell-background attribute to store table cell background color information.
+
 ## 25.2.1
 
 ### Patch Changes